no-auth asset links
new input field to add additional result fields for every asset URL with
a download link to the asset that does not require authorization for a
couple of hours.

affects the following actions:

- "Create Row" trigger
- "Update Row" trigger
- "Find Row" search

subsequent requests may fail, the additional URL is null then.

there are up to three new entries in the log:

1. timing - how many ms did the whole transaction take?
2. summary - how many urls were acquired, how many errors
3. (failures) - list details of all errors: url, path, api-url, response
   and exception (response or exception may be undefined)