add link-formula and button type to struct
testing for #1 revealed that whenever the square brackets were displayed
in a column description around the column type in create row and update
row, that they were not yet part of the known list.

during testing the old behavior, the following column types were
displayed in square brackets:

* link-formula
* button

adding the two to the list of column names as we know they are in use
(to hide) and were missing earlier.